feat(auction): paymentConsumer 수정
